auth

+++ Workout App/frontend/src/components/WorkoutDetails.js
@@ -1,28 +1,46 @@
-import { useWorkoutsContext } from "../hooks/useWorkoutContext"
-import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
-    const { dispatch } = useWorkoutsContext()
+  const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
-    const handleClick = async () => {
-        const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
-        })
-        const json = await response.json()
-        
-        if (response.ok) {
-            dispatch({type: 'DELETE_WORKOUT', payload: json })
-        }
+  const handleClick = async () => {
+    if (!user) {
+      return;
     }
-    return (
-        <div className="workout-details">
-            <h4>{workout.title}</h4>
-            <p><strong>Load (kg) : </strong>{workout.load}</p>
-            <p><strong>Reps : </strong>{workout.reps}</p>
-            <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-            <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
-        </div>
-    )
-}
+    const response = await fetch("/api/workouts/" + workout._id, {
+      method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
+    });
+    const json = await response.json();
 
-export default WorkoutDetails
+    if (response.ok) {
+      dispatch({ type: "DELETE_WORKOUT", payload: json });
+    }
+  };
+  return (
+    <div className="workout-details">
+      <h4>{workout.title}</h4>
+      <p>
+        <strong>Load (kg) : </strong>
+        {workout.load}
+      </p>
+      <p>
+        <strong>Reps : </strong>
+        {workout.reps}
+      </p>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
+      <span className="material-symbols-outlined" onClick={handleClick}>
+        delete
+      </span>
+    </div>
+  );
+};
+
+export default WorkoutDetails;
